fixed task8

+++ task8/Task8.java
@@ -2,7 +2,7 @@ public class Task8{
 
 	public static void main(String[] args){
 		int n = Integer.parseInt(args[0]);
-		int x = Integer.parseInt(args[1]);
+		float x = Float.parseFloat(args[1]);
 		if (args.length == 2){
 			if (n>0){
 				double result = 0.0d;
